Added in a fragment and viewmodel for a card. Not at all finished, but it's got most of the initial wiring set up in the view model for negotiating changing the icons when changes are made to the values of the card.

+++ app/src/main/java/org/ajar/umbrallegacy/content/AbilityDAO.kt
@@ -15,7 +15,7 @@ interface AbilityDAO {
     fun findByName(name: String): Ability?
 
     @Query("SELECT * FROM ${Ability.TABLE_NAME} WHERE id = :id")
-    fun findById(id: Long): Ability?
+    fun findById(id: Int): Ability?
 
     @Query("SELECT * FROM ${Ability.TABLE_NAME} order by name")
     fun selectAll(): Cursor
